requirements.txt
streamlit
nltk
newspaper3k
lxml_html_clean